filecopy improvements
